[Mv] Moving mute events in a specific class
